Fixed gem dependency issues
Still not sure as to why the dependencies declared in the gemspec aren't being included by default.
Ended up doing a manual require of these.

Installed the gem in a test Rails 3.1 project along with Rspec to ensure it worked correctly with no
dependency issues.

Updated the readme to include the render_views, otherwise response.body is empty.